Add solver auto-execution, on-chain executor, EIP-712 ecrecover, LP UI, metrics
- Solver worker: auto-registers default solver, processes pending intents,
  broadcasts settlements via WebSocket, updates solver stats
- On-chain executor: JSON-RPC balance checks, gas estimation, dry-run mode
  (set EXECUTOR_ENABLED=true + SOLVER_PRIVATE_KEY for live execution)
- EIP-712: full ecrecover via k256/secp256k1, domain/struct hash matches
  Settlement.sol exactly (ARI Exchange, correct field order)
- Frontend: Liquidity page with add/remove UI, CSS, navigation tab
- Frontend: EIP-712 field order fix in useSignIntent.ts
- Metrics endpoint (GET /v1/metrics): intent/solver/fill/position counts
- Enhanced health check: DB status, uptime, WS connections
- Settlement status endpoint (GET /v1/settlement/status)
- Subgraph API key support (SUBGRAPH_API_KEY env var)
- Clippy fixes across ari-engine, ari-solver, ari-gateway (0 warnings)
- All 46 Rust tests + 188 Solidity tests pass

+++ crates/ari-solver/src/auction.rs
@@ -109,8 +109,8 @@ impl DutchAuction {
         if elapsed >= self.duration {
             return self.end_price;
         }
-        let elapsed_ms = elapsed.as_millis() as u128;
-        let duration_ms = self.duration.as_millis() as u128;
+        let elapsed_ms = elapsed.as_millis();
+        let duration_ms = self.duration.as_millis();
         if duration_ms == 0 {
             return self.end_price;
         }
